create wrapper function runUpdate
