+++ lib/lib.js
@@ -1,6 +1,57 @@
 #!/usr/bin/env node
 
 var _ = require('underscore');
+var fs = require('fs');
+var request = require('request');
+
+
+
+// TODO: ditch 'file' input method, spoof api call instead
+// Make API call or read from file
+module.exports.getResponse = getResponse;
+function getResponse(method, address, callback) {
+    console.log('getting');
+    var response;
+    method = method || 'url';
+    address = address || 'https://pricing.us-east-1.amazonaws.com/offers/v1.0/aws/AmazonEC2/current/index.json';
+
+    // comment in to use saved api response for developement
+    method = 'file';
+    address =  './apiResponse.json';
+
+    if (method === 'url') {
+        // Get API response
+        request(address, function(err, response, body) {
+            if (err) {
+                throw new Error('bad request');
+            } else {
+                response = response.body;
+                return callback(null, response);
+            };
+        });
+    } else if (method === 'file') {
+        // read saved api response from file
+        fs.readFile(address, 'utf-8', function(err, buffer) {
+            console.log('reading');
+            if (err) {
+                throw new Error('bad request');
+            } else {
+                response = buffer;
+                console.log('returning');
+                return callback(null, response);
+            };
+        });
+    };
+};
+
+// main function; takes api response and writes mapping to file
+module.exports.createMapping = createMapping;
+function createMapping(response, output) {
+    var parsedResponse = parseResponse(response);
+    var mapping = formatMap(parsedResponse);
+
+    return mapping;
+};
 
 // Parses API response into legible object
 module.exports.parseResponse = parseResponse;
